test: move unit tests for ring buffer functionality to tests

+++ tests/ring_buf_size_max_test.c
@@ -4,7 +4,7 @@
 #include "ring_buf.h"
 
 int ring_buf_size_max_test(int argc, char **argv) {
-  assert(RING_BUF_SIZE_MAX == PTRDIFF_MAX + 1);
+  assert(RING_BUF_SIZE_MAX == (size_t)PTRDIFF_MAX + 1);
 
   return EXIT_SUCCESS;
 }
